Create StringCompression_KOWSIK-M.java
**String Compression** is a technique used to reduce the size of a string by replacing consecutive repeating characters with a single character followed by the count of its repetitions.

For example:

* Input: `"aaabbccddd"`
* Output: `"a3b2c2d3"`

This helps save space when strings have many repeating characters.

In Java, this can be implemented using a **loop** and a **StringBuilder** to count consecutive characters and append them efficiently. The algorithm ensures better performance by avoiding repeated string concatenations.

If the compressed string is not shorter than the original, the program returns the original string.